requirements.txt
# RL-A2A MVP Requirements
# Core Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
msgpack>=1.0.7
numpy>=1.24.0
pydantic>=2.0.0
requests>=2.31.0

# Security & Validation
pyjwt>=2.8.0
bcrypt>=4.0.0
slowapi>=0.1.9
limits>=3.5.0
bleach>=6.0.0
python-jose>=3.3.0
passlib>=1.7.4
cryptography>=41.0.0

# AI Providers
openai>=1.12.0
anthropic>=0.34.0
google-generativeai>=0.3.0
tiktoken>=0.5.0

# Local AI Support
ollama>=0.1.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
streamlit>=1.29.0
pandas>=2.1.0
seaborn>=0.12.0

# MCP Support
mcp>=1.0.0

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities  
python-dotenv>=1.0.0
redis>=5.0.0
aiofiles>=23.2.1
email-validator>=2.1.0
click>=8.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0